Set induction channel threshold to 5 instead of 3
This is a hack for the purposes of getting TP finding running
on the VD coldbox, where the collection/induction channel split does not match
ProtoDUNE. So there are some collection channels in the induction
processing stream and vice versa. Since we will only be using collection
channels for the current (2021-12) tests, we just set everything to
what we want for collection channels.